сценарий для getPictureElement перенесен в picture.js, удалена фунукция remove из gallery
